Python Code/requirements.txt
fastapi
uvicorn
geopandas
shapely
requests
python-dotenv